게임 승리 및 끝내기

+++ Assets/Scripts/Player.cs
@@ -68,7 +68,7 @@ public class Player : MonoBehaviour
     {
         if (!GameManager.instance.isLive) return;
 
-        GameManager.instance.health -= (int)Time.deltaTime * 10;
+        GameManager.instance.health -= Time.deltaTime * 10;
 
         if (GameManager.instance.health < 0)
         {
